refactor: update code style for index.js [HW-26]
