+++ JS/lesson-26/.gitignore
@@ -1,0 +1,3 @@
+# dependencies
+/node_modules
+/.parcel-cache
