+++ JS/lesson-26/index.js
@@ -1,23 +1,19 @@
-import {bigWrapper, wrapperRow, deleteAllButton, addTodoTextField, addTodoButton, transformedTodos} from './variables'
-import {getTodo, saveTodos, createTodo, deleteAll} from './funtions'
+import { bigWrapper, wrapperRow, deleteAllButton, addTodoTextField, addTodoButton } from "./components.js"
+import { getTodo, createTodo, deleteAll } from "./function.js"
 
 const root = document.getElementById('root')
 
 wrapperRow.append(deleteAllButton, addTodoTextField, addTodoButton)
 
-let todosFromStorage = localStorage.getItem('todos');
-const todosDB = todosFromStorage ? JSON.parse(todosFromStorage) : [];
 
-window.onclick = (event) => {
-  console.log('click')
-  saveTodos(todosDB);
-}
-
-const transformedTodos = todosDB.map(getTodo)
-
-bigWrapper.append(wrapperRow, ...transformedTodos)
+bigWrapper.append(wrapperRow)
 root.append(bigWrapper)
 
-addTodoButton.addEventListener('click', () => createTodo(todosDB))
+addTodoButton.addEventListener('click', createTodo)
 
 deleteAllButton.addEventListener('click', deleteAll)
+
+
+// ----------------- Render section -----------------
+
+// let todos = TodoController.generateTodos(2);
